Move Print Mode. Use default theme values.
