+++ src/components/deck/deck.js
@@ -23,9 +23,15 @@ import {
   printWrapperStyle
 } from './deck-styles';
 import { useAutoPlay } from '../../utils/use-auto-play';
+import defaultTheme from '../../theme/default-theme';
 
 export const DeckContext = createContext();
 const noop = () => {};
+
+/**
+ * The PDF DPI is 96. We want to scale the slide down because it's a 1:1 px to 1/100th of an inch.
+ * However there are some unchangeable margins that make 0.96 too big, so we use 0.959 to prevent overflow.
+ */
 const DEFAULT_PRINT_SCALE = 0.959;
 const DEFAULT_OVERVIEW_SCALE = 0.25;
 
@@ -64,8 +70,8 @@ const Deck = forwardRef(
       template,
       theme: {
         size: { width: nativeSlideWidth, height: nativeSlideHeight } = {
-          width: 1366,
-          height: 768
+          width: defaultTheme.size.width,
+          height: defaultTheme.size.height
         },
         Backdrop: UserProvidedBackdropComponent,
         backdropStyle: themeProvidedBackdropStyle = {
