+++ src/components/print-mode/index.js
@@ -1,8 +1,9 @@
 import React from 'react';
 import propTypes from 'prop-types';
 import styled, { createGlobalStyle } from 'styled-components';
-import Deck from '../components/deck/deck';
-import { AnimatedDiv } from '../components/slide/slide';
+import Deck from '../deck/deck';
+import { AnimatedDiv } from '../slide/slide';
+import defaultTheme from '../../theme/default-theme';
 
 const Backdrop = styled.div`
   background-color: white;
@@ -27,8 +28,8 @@ const PrintStyle = createGlobalStyle`
 
 export default function PrintMode(props) {
   const { children, theme, exportMode, pageSize, pageOrientation = '' } = props;
-  const width = theme?.size?.width || 1366;
-  const height = theme?.size?.height || 768;
+  const width = theme?.size?.width || defaultTheme.size.width;
+  const height = theme?.size?.height || defaultTheme.size.height;
   const computedPageSize = pageSize || `${width / 100}in ${height / 100}in`;
   return (
     <>
